+++ .env
@@ -1,0 +1,4 @@
+NODE_PATH=src
+REACT_APP_NAME=Imerso-Scan
+REACT_APP_DESCRIPTION=How many ways do you want to scan?
+REACT_APP_SOURCE_URL=/
