+++ src/App/TODO
@@ -1,0 +1,2 @@
+add redux
+    make call to mock api
