+++ src/components/ScanList/index.js
@@ -1,0 +1,1 @@
+export {default as ScanList} from './ScanList'
