+++ src/containers/ScanContainer/index.js
@@ -1,0 +1,1 @@
+export {default as ScanContainer} from './ScanContainer'
