+++ src/index.js
@@ -1,7 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
+import 'index.css';
+import App from 'App/App';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+const REACT_ROOT = document.getElementById('root')
+
+ReactDOM.render(<App />, REACT_ROOT);
